pyproject.toml
[project]
name = "AssetManager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.33.1",
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "websockets",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "pytest-asyncio>=1.3.0",
    "finance-datareader>=0.9.110",
]
